+++ lib/src/main/java/ai/freeplay/client/thin/internal/dto/TraceInfoDTO.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 import java.util.Map;
 import java.util.Objects;
+import java.util.UUID;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class TraceInfoDTO {
@@ -15,6 +16,7 @@ public class TraceInfoDTO {
     private Map<String, Object> customMetadata;
     private Map<String, Object> evalResults;
     private RecordDTO.TestRunInfoDTO testRunInfo;
+    private UUID parentId;
 
     @SuppressWarnings("unused")
     public TraceInfoDTO() {
@@ -27,8 +29,8 @@ public class TraceInfoDTO {
             String agentName,
             Map<String, Object> customMetadata,
             Map<String, Object> evalResults,
-            TestRunInfo testRunInfo
-
+            TestRunInfo testRunInfo,
+            UUID parentId
     ) {
         this.input = input;
         this.output = output;
@@ -36,6 +38,7 @@ public class TraceInfoDTO {
         this.customMetadata = customMetadata;
         this.evalResults = evalResults;
         this.testRunInfo = testRunInfo != null ? new RecordDTO.TestRunInfoDTO(testRunInfo.getTestRunId(), testRunInfo.getTestCaseId()) : null;
+        this.parentId = parentId;
     }
 
     @SuppressWarnings("unused")
@@ -68,16 +71,21 @@ public class TraceInfoDTO {
         return testRunInfo;
     }
 
+    @SuppressWarnings("unused")
+    public UUID getParentId() {
+        return parentId;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         TraceInfoDTO that = (TraceInfoDTO) o;
-        return Objects.equals(input, that.input) && Objects.equals(output, that.output) && Objects.equals(agentName, that.agentName) && Objects.equals(customMetadata, that.customMetadata) && Objects.equals(evalResults, that.evalResults);
+        return Objects.equals(input, that.input) && Objects.equals(output, that.output) && Objects.equals(agentName, that.agentName) && Objects.equals(customMetadata, that.customMetadata) && Objects.equals(evalResults, that.evalResults) && Objects.equals(parentId, that.parentId);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(input, output, agentName, customMetadata, evalResults);
+        return Objects.hash(input, output, agentName, customMetadata, evalResults, parentId);
     }
 }