+++ lib/src/main/java/ai/freeplay/client/thin/resources/sessions/TraceInfo.java
@@ -15,6 +15,7 @@ public class TraceInfo {
     public String output;
 
     public String agentName;
+    public UUID parentId;
 
     public Map<String, Object> customMetadata;
 
@@ -49,6 +50,13 @@ public class TraceInfo {
         return this;
     }
 
+    public TraceInfo parentId(
+            UUID parentId
+    ) {
+        this.parentId = parentId;
+        return this;
+    }
+
 
     public UUID getSessionId() {
         return sessionId;
@@ -76,6 +84,10 @@ public class TraceInfo {
 
     public Map<String, Object> getEvalResults() {
         return evalResults;
+    }
+
+    public UUID getParentId() {
+        return parentId;
     }
 
     @SuppressWarnings("unused")